infra/vpc/requirements.txt
attrs==20.3.0
aws-cdk.assets==1.107.0
aws-cdk.aws-cloudwatch==1.107.0
aws-cdk.aws-ec2==1.107.0
aws-cdk.aws-events==1.107.0
aws-cdk.aws-iam==1.107.0
aws-cdk.aws-kms==1.107.0
aws-cdk.aws-logs==1.107.0
aws-cdk.aws-s3==1.107.0
aws-cdk.aws-s3-assets==1.107.0
aws-cdk.aws-ssm==1.107.0
aws-cdk.cloud-assembly-schema==1.107.0
aws-cdk.core==1.107.0
aws-cdk.cx-api==1.107.0
aws-cdk.region-info==1.107.0
cattrs==1.5.0
constructs==3.3.75
jsii==1.30.0
publication==0.0.3
python-dateutil==2.8.1
six==1.16.0
typing-extensions==3.10.0.0
-e git+git@github.com:simonmackinnon/cloudguruchallenge-2021-06.git@9cbfd48c6dd33a3da1c63511a964f6919cfb9502#egg=vpc&subdirectory=infra/vpc
